backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
pybtex==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
alembic==1.12.1
python-dotenv==1.0.0
bibtexparser>=1.4.0
beautifulsoup4==4.12.2
requests==2.31.0
numpy

torch==2.8.0

accelerate==1.10.1
transformers==4.56.2

llama-cpp-python==0.2.90
huggingface-hub
openai
